correcting

+++ file_io/0-read_textfile.c
@@ -20,7 +20,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	buffer[letters + 1] = '\0';
 
 	err = read(f, buffer, letters);
-	if (err < letters)
+	if (err < (int) letters)
 	{
 		dprintf(STDERR_FILENO, "%s", buffer);
 		return (0);
